Make sure dir exists

+++ blueprints/dev-utils/_mac/macdylibbundler/macdylibbundler.py
@@ -19,7 +19,8 @@ class Package(MakeFilePackageBase):
 
     def fetch(self):
         if isinstance(self, GitSource):
-            utils.system(["git", "clean", "-xdf"], cwd=self.sourceDir())
+            if os.path.isdir(self.sourceDir()):
+                utils.system(["git", "clean", "-xdf"], cwd=self.sourceDir())
         return super().fetch()
 
     def install(self):
